pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-mobile-emulator-testing"
version = "1.0.0"
description = "Selenium test automation for Twitch mobile web using pytest"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Jorge Ribeiro", email = "mateusribeirojorge@gmail.com"}
]

dependencies = [
    "selenium>=4.26.0",
    "pytest>=8.3.0",
    "pytest-html>=4.1.0",
    "pytest-xdist>=3.6.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
# Test discovery
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Default options
addopts = [
    "-v",                           # Verbose (not -vv to reduce noise)
    "--html=reports/report.html",
    "--self-contained-html",
    "--strict-markers",
    "--Browser=Chrome",
]

# Test markers
markers = [
    "smoke: Quick sanity tests",
    "regression: Full test suite",
    "mobile: Mobile-specific tests",
    "streaming: Video streaming tests",
    "slow: Tests taking >30 seconds",
    "critical: Must-pass tests",
]

# Logging configuration
# Console logging - INFO level for terminal output
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s - %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

# File logging - DEBUG level for detailed logs
log_file = "reports/test_execution.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)8s] %(name)s:%(lineno)d - %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

log_auto_indent = true

# Ignore directories
norecursedirs = [".git", ".pytest_cache", "reports", "venv", "__pycache__"]
